requirements.txt
absl-py==2.2.0
aiofiles==24.1.0
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deepface==0.0.93
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.4.0
gdown==5.2.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
ImageHash==4.3.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.4.2
keras==2.12.0
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mtcnn==1.0.0
namex==0.0.8
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
Pillow==10.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
PyWavelets==1.8.0
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.9.4
rsa==4.9
scipy==1.15.2
six==1.17.0
soupsieve==2.6
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
